+++ HISTORY.md
@@ -5,7 +5,8 @@
 2. Fix builder call bug
 3. Add refresh projects button
 4. generic justfile
-5. Move refresh button to frm_project_versions
+5. Move refresh button to frm_project_usage
+6. Use revised psiutils.buttons
 
 ## Version 1.0.22 - 26 August 2026
 1. Use ModuleCallerBase from psiutils